microservices/image-conversion-service/requirements.txt
pillow==11.2.1
fastapi==0.104.1
uvicorn[standard]==0.24.0
aio-pika==9.3.0
httpx==0.25.1
python-dotenv==1.0.0
pydantic==2.4.2
cloudinary==1.36.0
requests==2.31.0
pillow-heif==0.18.0
opencv-python==4.8.1.78
imageio==2.34.0